Allow tests to report multiple diagnostics and run fixes for the first one
Because it cannot be seen which code changes belong to which diagnostic, a test with fixes for multiple diagnostics has never worked.

This change ensures that only the fixes for the first diagnostic are run. It allows for a test that contains multiple diagnostics, along with code changes for applying the first diagnostic. Such tests would ensure that fixers do not fix too much or are influenced by the existence of similar (but unrelated) diagnostics.